api_django/requirement.txt
django
djangorestframework
mysql-python
redis 
